pyproject.toml
[project]
name = "hls-frame-grabber"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.5.5",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "ffmpeg-python>=0.2.0",
    "httpx>=0.28.1",
    "matplotlib>=3.10.1",
    "numpy>=2.2.5",
    "pillow>=11.2.1",
    "psutil>=7.0.0",
    "pydantic>=2.11.4",
    "redis>=6.0.0",
    "requests>=2.32.3",
    "ultralytics>=8.3.128",
    "uvicorn>=0.34.2",
]
